pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "groq>=0.37.1",
    "langchain>=1.2.0",
    "langchain-groq>=1.1.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
